project_2/sql/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.3
databases==0.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.3.1
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3
